refactor(workspace): split @tool pimples into squad/workspace_tools.py; tools/workspace.py is pure impl
